pyproject.toml
[project]
name = "WxDatDecrypt"
version = "2.0.0"
description = "解密与查看微信 4.1 的图片，将微信缓存的 dat 文件解密为原始图片格式"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "hypercorn>=0.17.3",
    "pycryptodome>=3.23.0",
    "pymem>=1.14.0",
    "pywebview>=6.0",
    "rich>=14.2.0",
    "yara-python>=4.5.4",
]
